+++ src/test/java/co/comdor/github/LastMentionTestCase.java
@@ -23,10 +23,8 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package co.comdor.rest.model;
+package co.comdor.github;
 
-import co.comdor.github.LastMention;
-import co.comdor.github.Mention;
 import com.jcabi.github.*;
 import com.jcabi.github.mock.MkGithub;
 import com.jcabi.github.mock.MkStorage;
@@ -34,10 +32,9 @@ import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.Assert;
 import org.junit.Test;
-import javax.json.JsonObject;
 
 /**
- * Unit tests for {@link co.comdor.github.LastMention}
+ * Unit tests for {@link LastMention}
  * @author Mihai Andronache (amihaiemil@gmail.com)
  * @version $Id$
  * @since 0.0.1
